+++ src/manifolds/stiefel_manifold.jl
@@ -69,6 +69,7 @@ rgrad(Y, Δ)
 ```
 """
 function rgrad(Y::StiefelManifold, ∇L::AbstractMatrix)
+    ∇L = _match_backend(Y, ∇L) # TEMPORARY, see `_match_backend`
     ∇L - Y.A * (∇L' * Y.A)
 end